+++ misskey-api/src/endpoint/i/update.rs
@@ -57,6 +57,11 @@ pub struct Request {
     #[serde(skip_serializing_if = "Option::is_none")]
     #[builder(default, setter(strip_option))]
     pub hide_online_status: Option<bool>,
+    #[cfg(feature = "12-93-0")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "12-93-0")))]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(default, setter(strip_option))]
+    pub public_reactions: Option<bool>,
     #[serde(skip_serializing_if = "Option::is_none")]
     #[builder(default, setter(strip_option))]
     pub careful_bot: Option<bool>,
@@ -156,6 +161,8 @@ mod tests {
                 is_explorable: Some(false),
                 #[cfg(feature = "12-77-0")]
                 hide_online_status: Some(true),
+                #[cfg(feature = "12-93-0")]
+                public_reactions: Some(true),
                 careful_bot: Some(true),
                 auto_accept_followed: Some(true),
                 is_bot: Some(true),
@@ -210,6 +217,8 @@ mod tests {
                 is_explorable: None,
                 #[cfg(feature = "12-77-0")]
                 hide_online_status: None,
+                #[cfg(feature = "12-93-0")]
+                public_reactions: None,
                 careful_bot: None,
                 auto_accept_followed: None,
                 is_bot: None,