+++ misskey-api/src/endpoint/users.rs
@@ -44,6 +44,10 @@ pub mod gallery;
 #[cfg_attr(docsrs, doc(cfg(not(feature = "12-88-0"))))]
 pub mod recommendation;
 
+#[cfg(feature = "12-93-0")]
+#[cfg_attr(docsrs, doc(cfg(feature = "12-93-0")))]
+pub mod reactions;
+
 #[cfg(any(not(feature = "12-88-0"), feature = "12-89-0"))]
 #[derive(Serialize, PartialEq, Eq, Clone, Debug, Copy)]
 #[serde(rename_all = "camelCase")]